refactor: esm and jsx
