+++ package.json
@@ -7,17 +7,18 @@
   "keywords": [],
   "private": true,
   "packageManager": "pnpm@9.3.0",
-  "main": "dist/index.js",
+  "main": "dist/index.mjs",
+  "type": "module",
   "scripts": {
     "lint": "eslint --max-warnings=0 src",
     "lint:fix": "eslint --fix --max-warnings=0 src",
     "format": "prettier --write 'src/**/*.{js,jsx,ts,tsx,css,md,json}' --config ./.prettierrc",
-    "start": "NODE_ENV=production node dist/index.js",
-    "dev": "NODE_ENV=development tsx --watch ./src/index.ts",
+    "start": "NODE_ENV=production node dist/index.mjs",
+    "dev": "NODE_ENV=development tsx --watch ./src/index.mts",
     "test": "NODE_ENV=test vitest",
     "test:ci": "NODE_ENV=ci vitest run",
     "build:kill-dist": "rimraf ./dist",
-    "build:code": "tsc && tscpaths -p tsconfig.json -s ./src -o ./dist",
+    "build:code": "tsc && tsc-alias",
     "build": "pnpm run build:kill-dist && pnpm run build:code",
     "postbuild": "node postbuild.mjs",
     "db:explorer": "drizzle-kit studio",
@@ -35,7 +36,7 @@
     "pg": "^8.12.0",
     "prettier": "^3.3.2",
     "rimraf": "^5.0.7",
-    "tscpaths": "^0.0.9",
+    "tsc-alias": "^1.8.10",
     "tsx": "^4.15.7",
     "typescript": "^5.5.2",
     "vite-tsconfig-paths": "^4.3.2",
