+++ src/config/db/index.mts
@@ -1,0 +1,1 @@
+export { db } from "./drizzle.mjs";
