--- src/config/db/index.ts
@@ -1,1 +1,0 @@
-export { db } from "./drizzle";
