+++ src/routers/app/ui/index.tsx
@@ -1,10 +1,28 @@
-import type { ServerContext } from "@/types/hono";
+import type { ServerContext } from "@/types/hono.mjs";
 import { Hono } from "hono";
 import { html } from "hono/html";
+import type { FC } from "hono/jsx";
 
-import { db } from "@/config/db";
+import { db } from "@/config/db/index.mjs";
 
 const app = new Hono<ServerContext>();
+
+const Layout: FC = ({ children }) => {
+  return (
+    <html>
+      <head>
+        <script
+          src="https://unpkg.com/htmx.org@1.9.10"
+          integrity="sha384-D1Kt99CQMDuVetoL1lrYwg5t+9QdHe7NLX/SoJYkXDFfX37iInKRy5xLSi8nO7UC"
+          crossorigin="anonymous"
+        ></script>
+        <script src="https://unpkg.com/htmx.org/dist/ext/json-enc.js"></script>
+        <script src="https://cdn.tailwindcss.com"></script>
+      </head>
+      <body>{children}</body>
+    </html>
+  );
+};
 
 app.get("/login", async (c) => {
   const user = c.var.user;
@@ -13,11 +31,13 @@ app.get("/login", async (c) => {
     return c.redirect("/app");
   }
 
-  return c.render(
-    html`<div style="display: grid;gap: 0.5rem;">
-      <a href="/app/login/github">Login with Github</a>
-      <a href="/app">← Go back</a>
-    </div>`,
+  return c.html(
+    <Layout>
+      <div style="display: grid;gap: 0.5rem;">
+        <a href="/app/login/github">Login with Github</a>
+        <a href="/app">← Go back</a>
+      </div>
+    </Layout>,
   );
 });
