+++ src/types/hono.mts
@@ -1,7 +1,7 @@
 import type { HttpBindings } from "@hono/node-server";
 import type { Session, User } from "lucia";
 
-import type { services as servicesTable } from "@/config/db/schema";
+import type { services as servicesTable } from "@/config/db/schema.mjs";
 type Service = typeof servicesTable.$inferSelect;
 
 type Bindings = HttpBindings & {};
